Key of cards changed to long

+++ app/src/main/java/com/kai/unogame/listener/DeckCardsListener.java
@@ -3,6 +3,6 @@ package com.kai.unogame.listener;
 import java.util.ArrayList;
 
 public interface DeckCardsListener {
-    void deckCardsSuccess(ArrayList<Integer> list);
+    void deckCardsSuccess(ArrayList<Long> list);
     void deckFailure(String message);
 }
