+++ app/src/main/java/com/kai/unogame/listener/UserCardsListener.java
@@ -3,6 +3,6 @@ package com.kai.unogame.listener;
 import java.util.ArrayList;
 
 public interface UserCardsListener {
-    void userCardsSuccess(ArrayList<Integer> list);
+    void userCardsSuccess(ArrayList<Long> list);
     void userFailure(String message);
 }
